requirements.txt
asgiref==3.2.7
astroid==2.4.2
boto3==1.14.42
botocore==1.17.42
Django==3.0.7
django-extensions==3.0.5
django-s3direct==1.1.8
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-datatables==0.5.2
docutils==0.15.2
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.780
mypy-extensions==0.4.3
pep257==0.7.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
wrapt==1.12.1
